+++ src/main/java/chylex/hee/block/BlockInstabilityOrbOre.java
@@ -28,8 +28,8 @@ public class BlockInstabilityOrbOre extends BlockAbstractOre{
 	@Override
 	protected void onOreMined(EntityPlayer player, ArrayList<ItemStack> drops, int x, int y, int z, int meta, int fortune){
 		if (KnowledgeRegistrations.INSTABILITY_ORB_ORE.tryUnlockFragment(player,0.12F).stopTrying)return;
-		if (KnowledgeRegistrations.INSTABILITY_ORB.tryUnlockFragment(player,0.07F,new short[]{ 0,2 }).stopTrying)return;
-		if (player.worldObj.rand.nextInt(3) != 0)KnowledgeRegistrations.ENHANCED_BREWING_STAND.tryUnlockFragment(player,0.1F,new short[]{ 3 });
-		else KnowledgeRegistrations.ENERGY_EXTRACTION_TABLE.tryUnlockFragment(player,0.15F,new short[]{ 6 });
+		if (KnowledgeRegistrations.INSTABILITY_ORB.tryUnlockFragment(player,0.07F,new byte[]{ 0,2 }).stopTrying)return;
+		if (player.worldObj.rand.nextInt(3) != 0)KnowledgeRegistrations.ENHANCED_BREWING_STAND.tryUnlockFragment(player,0.1F,new byte[]{ 3 });
+		else KnowledgeRegistrations.ENERGY_EXTRACTION_TABLE.tryUnlockFragment(player,0.15F,new byte[]{ 6 });
 	}
 }
