+++ src/main/java/chylex/hee/entity/boss/EntityMiniBossFireFiend.java
@@ -122,7 +122,7 @@ public class EntityMiniBossFireFiend extends EntityFlying implements IBossDispla
 				motionY += 0.4D;
 			}
 			
-			for(EntityPlayer observer:ObservationUtil.getAllObservers(this,20D))KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment(observer,1F,new short[]{ 0,1 });
+			for(EntityPlayer observer:ObservationUtil.getAllObservers(this,20D))KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment(observer,1F,new byte[]{ 0,1 });
 		}
 	}
 	
@@ -160,7 +160,7 @@ public class EntityMiniBossFireFiend extends EntityFlying implements IBossDispla
 				worldObj.spawnEntityInWorld(new EntityProjectileGolemFireball(worldObj,this,posX,posY-0.2D,posZ,target.posX-posX,target.posY-posY,target.posZ-posZ));
 				
 				if (rand.nextBoolean() && rand.nextBoolean()){
-					for(EntityPlayer observer:ObservationUtil.getAllObservers(this,60D))KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment(observer,0.67F,new short[]{ 0,1,2 });
+					for(EntityPlayer observer:ObservationUtil.getAllObservers(this,60D))KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment(observer,0.67F,new byte[]{ 0,1,2 });
 				}
 			}
 
@@ -207,7 +207,7 @@ public class EntityMiniBossFireFiend extends EntityFlying implements IBossDispla
 				if (--fireballAttackTimer < 0)attackStage = STAGE_SHOOTING;
 				
 				if (rand.nextInt(18) == 0){
-					for(EntityPlayer observer:ObservationUtil.getAllObservers(this,60D))KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment(observer,0.15F,new short[]{ 0,1,2,3 });
+					for(EntityPlayer observer:ObservationUtil.getAllObservers(this,60D))KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment(observer,0.15F,new byte[]{ 0,1,2,3 });
 				}
 			}
 			else if (attackStage == STAGE_SHOOTING){
@@ -220,7 +220,7 @@ public class EntityMiniBossFireFiend extends EntityFlying implements IBossDispla
 						worldObj.spawnEntityInWorld(new EntityProjectileGolemFireball(worldObj,this,ballX,ballY,ballZ,target.posX-ballX,target.boundingBox.minY+target.height*0.5F-ballY,target.posZ-ballZ));
 						
 						if (rand.nextInt(10) == 0){
-							for(EntityPlayer observer:ObservationUtil.getAllObservers(this,60D))KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment(observer,0.3F,new short[]{ 0,1,2,3,4 });
+							for(EntityPlayer observer:ObservationUtil.getAllObservers(this,60D))KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment(observer,0.3F,new byte[]{ 0,1,2,3,4 });
 						}
 						
 						iter.remove();
@@ -256,7 +256,7 @@ public class EntityMiniBossFireFiend extends EntityFlying implements IBossDispla
 			++damageInflicted;
 			
 			if (e instanceof EntityPlayer && rand.nextInt(15) == 0){
-				KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment((EntityPlayer)e,0.1F,new short[]{ 0,1,2,3,4 });
+				KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment((EntityPlayer)e,0.1F,new byte[]{ 0,1,2,3,4 });
 			}
 		}
 		
@@ -289,7 +289,7 @@ public class EntityMiniBossFireFiend extends EntityFlying implements IBossDispla
 	protected void dropFewItems(boolean recentlyHit, int looting){
 		for(int a = 0; a < 60; a++)entityDropItem(new ItemStack(ItemList.essence,2,EssenceType.FIERY.getItemDamage()),rand.nextFloat()*height);
 		
-		for(EntityPlayer observer:ObservationUtil.getAllObservers(this,60D))KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment(observer,1F,new short[]{ 5,6 });
+		for(EntityPlayer observer:ObservationUtil.getAllObservers(this,60D))KnowledgeRegistrations.FIRE_FIEND.tryUnlockFragment(observer,1F,new byte[]{ 5,6 });
 	}
 	
 	@Override
